requirements.txt
# Основные зависимости Django
asgiref==3.8.1
Django==5.0.14
Pillow==10.3.0
sqlparse==0.5.0
django-braces==1.15.0
django-embed-video==1.4.9
pymemcache==4.0.0
django-debug-toolbar==4.3.0
redis==5.0.4
django-redisboard==8.4.0
djangorestframework==3.15.1
requests==2.31.0

# Для WebSocket чата
channels[daphne]==4.1.0
channels-redis==4.2.0
daphne==4.2.1
autobahn==25.12.1
twisted==25.5.0

# База данных
psycopg2-binary==2.9.9

# Для продакшена (опционально)
uwsgi==2.0.25.1
python-decouple==3.8

# Для Telegram бота (если нужен)
aiogram==3.23.0
aiohttp==3.13.2
python-telegram-bot==22.5




django-tinymce==4.0.0